refactor(UserService): 인터페이스 분리 및 데코레이터 패턴 적용
