+++ app/src/main/java/camp/nextstep/controller/LoginController.java
@@ -1,7 +1,7 @@
 package camp.nextstep.controller;
 
 import camp.nextstep.domain.User;
-import camp.nextstep.service.UserService;
+import camp.nextstep.service.TxUserService;
 import com.interface21.beans.factory.annotation.Autowired;
 import com.interface21.context.stereotype.Controller;
 import com.interface21.web.bind.annotation.RequestMapping;
@@ -18,10 +18,11 @@ public class LoginController {
 
     private static final Logger log = LoggerFactory.getLogger(LoginController.class);
 
-    private final UserService userService;
+    private final TxUserService userService;
 
     @Autowired
-    public LoginController(final UserService userService) {
+    public LoginController(final TxUserService userService) {
+        System.out.println("LoginController 생성 : userService = " + userService.getClass());
         this.userService = userService;
     }
