+++ jdbc/src/main/java/camp/nextstep/transaction/TransactionHandler.java
@@ -2,7 +2,7 @@ package camp.nextstep.transaction;
 
 import camp.nextstep.jdbc.CannotCloseJdbcConnectionException;
 import camp.nextstep.jdbc.CannotGetJdbcConnectionException;
-import camp.nextstep.jdbc.datasource.ConnectionUtils;
+import camp.nextstep.jdbc.datasource.DataSourceUtils;
 import camp.nextstep.transaction.support.TransactionSynchronizationManager;
 import java.sql.Connection;
 import java.sql.SQLException;
@@ -11,7 +11,7 @@ import javax.sql.DataSource;
 public class TransactionHandler {
     public void begin(DataSource dataSource) {
         try {
-            Connection connection = ConnectionUtils.getConnection(dataSource);
+            Connection connection = DataSourceUtils.getConnection(dataSource);
             connection.setAutoCommit(false);
         } catch (SQLException e) {
             throw new CannotGetJdbcConnectionException("JDBC 커넥션 획득 실패", e);
@@ -21,7 +21,7 @@ public class TransactionHandler {
     public void commit(DataSource dataSource) {
         Connection connection = null;
         try {
-            connection = ConnectionUtils.getConnection(dataSource);
+            connection = DataSourceUtils.getConnection(dataSource);
             connection.commit();
         } catch (SQLException e) {
             throw new CannotCommitException("커밋 실패", e);
@@ -34,7 +34,7 @@ public class TransactionHandler {
     public void rollback(DataSource dataSource) {
         Connection connection = null;
         try {
-            connection = ConnectionUtils.getConnection(dataSource);
+            connection = DataSourceUtils.getConnection(dataSource);
             connection.rollback();
         } catch (SQLException e) {
             throw new CannotRollbackException("롤백 실패", e);
